Rename files and add `big` and `borderRadius` props
