+++ assets/javascripts/kitten/components/expandable/expand-board/examples.js
@@ -4,7 +4,7 @@ import { ExpandBoard } from 'kitten/components/expandable/expand-board'
 import { List } from 'kitten/components/lists/list'
 import { Grid, GridCol } from 'kitten/components/grid/grid'
 import { Text } from 'kitten/components/typography/text'
-import { pxToRem } from '../../helpers/utils/typography'
+import { pxToRem } from '../../../helpers/utils/typography'
 
 const fadeInAnimation = () =>
   keyframes`
