+++ assets/javascripts/kitten/components/expandable/expand-board/test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import { shallow } from 'enzyme'
-import { ExpandBoard } from '../../components/expandable/expand-board'
-import { Button } from '../../components/buttons/button'
+import { ExpandBoard } from '../../../components/expandable/expand-board'
+import { Button } from '../../../components/buttons/button'
 
 describe('<ExpandBoard />', () => {
   describe('with <ExpandBoard.Button> and <ExpandBoard.Content>', () => {
